+++ lib/base/base_state_status_page.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+
+import 'base_controller.dart';
+import 'base_model.dart';
+import 'base_page.dart';
+
+// ignore: must_be_immutable
+abstract class BaseStatePage<T extends BaseStatusController,
+    M extends BaseModel> extends BaseStatusPage<T, M> {
+  BaseStatePage({super.key});
+
+  late final T initController;
+
+  @override
+  T get binding => initController;
+
+  @override
+  T get controller => initController;
+
+  @override
+  Widget build(BuildContext context) {
+    return GetBuilder<T>(
+      init: initController,
+      builder: (controller) {
+        if (!enableWillPop) return scaffold;
+        return WillPopScope(
+          onWillPop: controller.onWillPop,
+          child: scaffold,
+        );
+      },
+    );
+  }
+}